+++ client/src/extension.ts
@@ -1,6 +1,6 @@
 
 import * as path from 'path';
-import { ExtensionContext } from 'vscode';
+import { ExtensionContext, commands, Selection, TextEditor } from 'vscode';
 
 import {
 	LanguageClient,
@@ -11,7 +11,55 @@ import {
 
 let client: LanguageClient;
 
+/**
+ * Surround the current selection with bbcode delimiters like [i] and [/i].
+ * @param editor Current editor.
+ * @param delimitCharacters Characters inside the bbcode [], such as "i" or "b"
+ */
+function bbcodeDelimit(editor: TextEditor, delimitCharacters: string) {
+	if (!editor) {
+		return;
+	}
+
+	let open = `[${delimitCharacters}]`;
+	let close = `[/${delimitCharacters}]`
+	let document = editor.document;
+	let selection = editor.selection;
+	if (selection.isEmpty) {
+		editor.edit(edit => edit.insert(selection.start, `${open}${close}`)).then(x => {
+			let newStartIndex = document.offsetAt(editor.selection.start) - close.length;
+			editor.selection = new Selection(
+				document.positionAt(newStartIndex),
+				document.positionAt(newStartIndex)
+			);
+		});
+	}
+	else {
+		let word = document.getText(selection);
+		editor.edit(edit => edit.replace(selection, `${open}${word}${close}`)).then(x => {
+			let newStartIndex = document.offsetAt(editor.selection.start) + open.length;
+			let newEndIndex = document.offsetAt(editor.selection.end) - close.length;
+			editor.selection = new Selection(
+				document.positionAt(newStartIndex),
+				document.positionAt(newEndIndex)
+			);
+		});
+	}
+}
+
 export function activate(context: ExtensionContext) {
+	// Register our commands
+	const italicsCommandRegistration = commands.registerTextEditorCommand(
+		'choicescript.italicize', (editor) => {
+			bbcodeDelimit(editor, "i");
+	});
+	const boldCommandRegistration = commands.registerTextEditorCommand(
+		'choicescript.bold', (editor) => {
+			bbcodeDelimit(editor, "b");
+	});
+
+	context.subscriptions.push(italicsCommandRegistration, boldCommandRegistration);
+
 	// The server is implemented in node
 	let serverModule = context.asAbsolutePath(
 		path.join('server', 'out', 'server.js')